backend/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
backoff==2.2.1
boto3==1.35.21
botocore==1.35.21
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.5
mangum==0.17.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.20.0
ruff==0.6.2
s3transfer==0.10.2
schema==0.7.7
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
starlette==0.37.2
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
websockets==12.0
